chore: renamed dir name actions to action
