+++ components/molecules/action/ActionRateSlider/ActionRateSlider.module.scss
