+++ components/molecules/action/ActionRateSlider/ActionRateSlider.stories.tsx
