+++ pages/category/[category]/action.tsx
@@ -4,10 +4,10 @@ import { GetStaticPaths, GetStaticProps, NextPage } from 'next'
 import { useRouter } from 'next/router'
 import { Box, Spinner } from '@chakra-ui/react'
 import DatasourceFooter from 'components/DatasourceFooter'
-import ActionCompleteBtn from 'components/molecules/actions/ActionCompleteBtn/ActionCompleteBtn'
-import ActionHeader from 'components/molecules/actions/ActionHeader/ActionHeader'
-import ActionChangeRateDialog from 'components/organisms/actions/ActionChangeRateDialog/ActionChangeRateDialog'
-import ActionItem from 'components/organisms/actions/ActionItem/ActionItem'
+import ActionCompleteBtn from 'components/molecules/action/ActionCompleteBtn/ActionCompleteBtn'
+import ActionHeader from 'components/molecules/action/ActionHeader/ActionHeader'
+import ActionChangeRateDialog from 'components/organisms/action/ActionChangeRateDialog/ActionChangeRateDialog'
+import ActionItem from 'components/organisms/action/ActionItem/ActionItem'
 import QuestionContainer from 'components/organisms/questions/Container'
 import { useActions } from 'hooks/actions'
 import { useProfile } from 'hooks/profile'
